Split views.py into views/, remove filters.py
