--- publish/filters.py
@@ -1,11 +1,0 @@
-from django_filters import rest_framework as filters
-
-from .models import Asset
-
-
-class AssetFilter(filters.FilterSet):
-    path_prefix = filters.CharFilter(lookup_expr='startswith', field_name='path')
-
-    class Meta:
-        model = Asset
-        fields = ['path', 'path_prefix']
