+++ publish/views/__init__.py
@@ -1,0 +1,3 @@
+from .asset import AssetViewSet
+from .dandiset import DandisetViewSet
+from .version import VersionViewSet
