+++ publish/views/asset.py
@@ -1,0 +1,32 @@
+from django_filters import rest_framework as filters
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.viewsets import ReadOnlyModelViewSet
+from rest_framework_extensions.mixins import DetailSerializerMixin, NestedViewSetMixin
+
+from publish.models import Asset
+from publish.serializers import AssetDetailSerializer, AssetSerializer
+from publish.views.common import DandiPagination
+
+
+class AssetFilter(filters.FilterSet):
+    path_prefix = filters.CharFilter(lookup_expr='startswith', field_name='path')
+
+    class Meta:
+        model = Asset
+        fields = ['path_prefix']
+
+
+class AssetViewSet(NestedViewSetMixin, DetailSerializerMixin, ReadOnlyModelViewSet):
+    queryset = Asset.objects.all().select_related('version__dandiset')
+    queryset_detail = queryset
+
+    permission_classes = [IsAuthenticatedOrReadOnly]
+    serializer_class = AssetSerializer
+    serializer_detail_class = AssetDetailSerializer
+    pagination_class = DandiPagination
+
+    lookup_field = 'uuid'
+    lookup_value_regex = Asset.UUID_REGEX
+
+    filter_backends = [filters.DjangoFilterBackend]
+    filterset_class = AssetFilter
