+++ publish/views/common.py
@@ -1,0 +1,7 @@
+from rest_framework.pagination import PageNumberPagination
+
+
+class DandiPagination(PageNumberPagination):
+    page_size = 25
+    max_page_size = 100
+    page_size_query_param = 'page_size'
